# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer API Documentation",
        "description": "Documentation de l'API Customer",
        "contact": {
            "email": "support@example.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/customer-contact-titles": {
            "get": {
                "tags": [
                    "CustomerContactTitles"
                ],
                "summary": "Liste toutes les civilités clients",
                "operationId": "773d77e6b50e7041252de8fb678c1afd",
                "responses": {
                    "200": {
                        "description": "Liste des civilités clients récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CustomerContactTitle"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des civilités clients récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CustomerContactTitles"
                ],
                "summary": "Crée une nouvelle civilité client",
                "operationId": "84f87c10072b162f180b0ee81525dec2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCustomerContactTitleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Civilité client créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerContactTitle"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Civilité client créée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/customer-contact-titles/{id}": {
            "get": {
                "tags": [
                    "CustomerContactTitles"
                ],
                "summary": "Affiche les détails d'une civilité client spécifique",
                "operationId": "67955bc90bc8801b2aa88513fda60cc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la civilité client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Civilité client récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerContactTitle"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Civilité client récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Civilité client non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "CustomerContactTitles"
                ],
                "summary": "Met à jour une civilité client existante",
                "operationId": "07f418e20ed5d548288009f1c0709552",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la civilité client à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomerContactTitleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Civilité client mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerContactTitle"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Civilité client mise à jour avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Civilité client non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CustomerContactTitles"
                ],
                "summary": "Supprime une civilité client",
                "operationId": "3ea493d84435d748a54eb37bdf79b942",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la civilité client à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Civilité client supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Civilité client supprimée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Civilité client non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCustomerContactTitleRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Nom de la civilité",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Dr"
                    }
                },
                "type": "object"
            },
            "UpdateCustomerContactTitleRequest": {
                "properties": {
                    "name": {
                        "description": "Nom de la civilité",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Dr"
                    }
                },
                "type": "object"
            },
...
            "CustomerContactTitle": {
                "title": "CustomerContactTitle",
                "description": "Modèle de civilité client",
                "properties": {
                    "id_customer_contact_title": {
                        "description": "ID unique de la civilité client",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de la civilité",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Mr"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière modification",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}